# contracts/oar-openapi.yaml
openapi: 3.1.0
info:
  title: Organization Autorunner HTTP API
  version: 0.3.0
  description: |
    Canonical workspace HTTP contract for Organization Autorunner.

    Topics and cards are the primary durable resources. Threads remain backing
    infrastructure surfaced through read-only inspection and workspace views.
servers:
  - url: http://127.0.0.1:8000
...
paths:
...
  /threads:
    get:
      operationId: listThreads
      summary: List backing threads
      tags: [threads]
      x-oar-surface: canonical
      x-oar-command-id: threads.list
      x-oar-cli-path: threads list
      x-oar-why: Inspect backing infrastructure threads without making them the primary planning noun.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ threads }`."
      x-oar-error-codes: [auth_required, invalid_token]
      x-oar-concepts: [threads, inspection]
      x-oar-stability: beta
      responses:
        '200':
          description: Thread list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadsResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
...
  /cards:
    get:
      operationId: listCards
      summary: List cards
      tags: [cards]
      x-oar-surface: canonical
      x-oar-command-id: cards.list
      x-oar-cli-path: cards list
      x-oar-why: Scan first-class card resources across boards.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ cards }`."
      x-oar-error-codes: [auth_required, invalid_token]
      x-oar-concepts: [cards]
      x-oar-stability: beta
      parameters:
        - name: archived_only
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: When true, return only archived (soft-deleted) cards.
        - name: tombstoned_only
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Trash-list alias for archived_only (cards are archived, not tombstoned).
      responses:
        '200':
          description: Card list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardsResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
...
  /events:
    get:
      operationId: listEvents
      summary: List events
      tags: [events]
      x-oar-surface: canonical
      x-oar-command-id: events.list
      x-oar-cli-path: events list
      x-oar-why: Inspect append-only event history across the workspace.
      x-oar-input-mode: none
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ events }`."
      x-oar-error-codes: [auth_required, invalid_token]
      x-oar-concepts: [events]
      x-oar-stability: beta
      responses:
        '200':
          description: Event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: createEvent
      summary: Create event
      tags: [events]
      x-oar-surface: canonical
      x-oar-command-id: events.create
      x-oar-cli-path: events create
      x-oar-why: Append an event that links first-class resources and evidence through typed refs.
      x-oar-input-mode: json-body
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ event }`."
      x-oar-error-codes: [auth_required, invalid_request, invalid_token]
      x-oar-concepts: [events, write]
      x-oar-stability: beta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  responses:
    ErrorResponse:
      description: Error payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
      required: [code, message]
...
    TypedRef:
      type: string
      description: Typed ref string such as `topic:roadmap` or `artifact:art_123`.
    Provenance:
      type: object
      properties:
        sources:
          type: array
          items: { type: string }
        notes: { type: string }
        by_field:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
      required: [sources]
...
    Event:
      type: object
      properties:
        id: { type: string }
        ts: { type: string, format: date-time }
        type: { type: string }
        actor_id: { type: string }
        thread_ref: { $ref: '#/components/schemas/TypedRef' }
        refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        summary: { type: string }
        payload:
          type: object
          additionalProperties: true
        provenance: { $ref: '#/components/schemas/Provenance' }
      required: [id, ts, type, actor_id, refs, summary, provenance]
...
    Thread:
      type: object
      properties:
        id: { type: string }
        topic_ref: { $ref: '#/components/schemas/TypedRef' }
        title: { type: string }
        status:
          type: string
          enum: [active, archived]
        priority:
          type: string
          enum: [p0, p1, p2, p3]
        tags:
          type: array
          items: { type: string }
        latest_event_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
        updated_at: { type: string, format: date-time }
        updated_by: { type: string }
        provenance: { $ref: '#/components/schemas/Provenance' }
      required: [id, title, status, priority, tags, created_at, created_by, updated_at, updated_by, provenance]
...
    Card:
      type: object
      properties:
        id: { type: string }
        board_ref: { $ref: '#/components/schemas/TypedRef' }
        topic_ref: { $ref: '#/components/schemas/TypedRef' }
        thread_ref: { $ref: '#/components/schemas/TypedRef' }
        document_ref: { $ref: '#/components/schemas/TypedRef' }
        title: { type: string }
        summary: { type: string }
        column_key:
          type: string
          enum: [backlog, ready, in_progress, blocked, review, done]
        rank: { type: string }
        assignee_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        risk:
          type: string
          enum: [low, medium, high, critical]
        resolution:
          type: string
          enum: [unresolved, completed, canceled, superseded]
        resolution_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        related_refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        created_at: { type: string, format: date-time }
        created_by: { type: string }
        updated_at: { type: string, format: date-time }
        updated_by: { type: string }
        provenance: { $ref: '#/components/schemas/Provenance' }
      required:
        [id, board_ref, title, summary, column_key, rank, assignee_refs, risk, resolution, resolution_refs,
         related_refs, created_at, created_by, updated_at, updated_by, provenance]
...
    EventCreateInput:
      type: object
      properties:
        type: { type: string }
        actor_id: { type: string }
        thread_ref: { $ref: '#/components/schemas/TypedRef' }
        refs:
          type: array
          items: { $ref: '#/components/schemas/TypedRef' }
        summary: { type: string }
        payload:
          type: object
          additionalProperties: true
        provenance: { $ref: '#/components/schemas/Provenance' }
      required: [type, actor_id, refs, summary, provenance]
...
    CreateEventRequest:
      type: object
      properties:
        event: { $ref: '#/components/schemas/EventCreateInput' }
      required: [event]
...
    ThreadsResponse:
      type: object
      properties:
        threads:
          type: array
          items: { $ref: '#/components/schemas/Thread' }
      required: [threads]
...
    CardsResponse:
      type: object
      properties:
        cards:
          type: array
          items: { $ref: '#/components/schemas/Card' }
      required: [cards]
...
    EventResponse:
      type: object
      properties:
        event: { $ref: '#/components/schemas/Event' }
      required: [event]
    EventsResponse:
      type: object
      properties:
        events:
          type: array
          items: { $ref: '#/components/schemas/Event' }
      required: [events]
...